requirements.txt
anthropic
openai
google-cloud-texttospeech
google-auth
google-api-python-client
google-auth-httplib2
fal-client
requests
python-dotenv
openai-whisper
edge-tts
flask
schedule
feedparser
wikipedia-api